move mailer to mail namespace
